+++ WebAPI/Reflection/KalturaJsonDeserializer.cs
@@ -1719,7 +1719,7 @@ namespace WebAPI.Models.ConditionalAccess
             {
                 if (parameters.ContainsKey("adapterId") && parameters["adapterId"] != null)
                 {
-                    AdapterId = (Int64) Convert.ChangeType(parameters["adapterId"], typeof(Int64));
+                    AdapterId = (Int32) Convert.ChangeType(parameters["adapterId"], typeof(Int32));
                 }
             }
         }
